requirements.txt
pdfminer.six
pytest
